fusionlab-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Data Validation
pydantic

# File Handling
python-multipart==0.0.6
Pillow==10.1.0
aiofiles==23.2.1

# Configuration
python-dotenv

# Development Tools
ruff
pytest
httpx